requirements.txt
allure-python-commons==2.13.5
allure-pytest==2.13.5
pluggy==1.5.0
pytest==8.3.3
selenium==4.25.0